Starting to break components into Presentational & Container directories
