--- client/src/components/LandingPage/index.js
@@ -1,14 +1,0 @@
-import React from "react"
-import styled from "styled-components"
-
-const LandingPageStyles = styled.div``
-
-const LandingPage = props => {
-  return (
-    <LandingPageStyles>
-      <div>Landing Page</div>
-    </LandingPageStyles>
-  )
-}
-
-export default LandingPage
