+++ client/src/components/Pages/LandingPage/LandingPageNav.js
@@ -1,0 +1,8 @@
+import React from "react"
+import LandingNav from "../../Presentational/LandingNav"
+
+const LandingPageNav = () => {
+  return <LandingNav />
+}
+
+export default LandingPageNav
