+++ client/src/components/Pages/LandingPage/index.js
@@ -1,0 +1,8 @@
+import React from "react"
+import Landing from "../../Presentational/Landing"
+
+const LandingPage = props => {
+  return <Landing />
+}
+
+export default LandingPage
