+++ client/src/components/Presentational/Landing/index.js
@@ -1,0 +1,12 @@
+import React from "react"
+import * as s from "./styles"
+
+const Landing = props => {
+  return (
+    <s.LandingStyles>
+      <div> LANDING PAGE </div>
+    </s.LandingStyles>
+  )
+}
+
+export default Landing
