+++ client/src/components/Presentational/Landing/styles.js
@@ -1,0 +1,3 @@
+import styled from "styled-components"
+
+export const LandingStyles = styled.div``
